playback/projectscene: extract VolumePressureMeterItemBase
