+++ src/playback/CMakeLists.txt
@@ -53,6 +53,8 @@ set(MODULE_SRC
     ${CMAKE_CURRENT_LIST_DIR}/view/common/playbackmetermodel.h
     ${CMAKE_CURRENT_LIST_DIR}/view/common/metermodel.cpp
     ${CMAKE_CURRENT_LIST_DIR}/view/common/metermodel.h
+    ${CMAKE_CURRENT_LIST_DIR}/view/common/volumepressuremeteritembase.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/view/common/volumepressuremeteritembase.h
     ${CMAKE_CURRENT_LIST_DIR}/view/common/horizontalvolumepressuremeteritem.cpp
     ${CMAKE_CURRENT_LIST_DIR}/view/common/horizontalvolumepressuremeteritem.h
